+++ examples/working/misc/cond_ddf/include/cond.h
@@ -1,0 +1,10 @@
+#ifndef _COND_H
+#define _COND_H
+
+#include "hocl.h"
+
+void is_even(IN int *i, OUT bool *o);
+void incr(IN int *i, OUT int *o);
+void mult2(IN int *i, OUT int *o);
+
+#endif