+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * The so-called `prelude` file is now called the _standard library_ (`lib/hocl/stdlib.hcl`). This
   file is now automatically parsed by the compiler, unless the `-no_stdlib` argument is given
 * Increased the number of SystemC examples 
+* `switch` and `merge` dedicated actors for expressing DDF-style conditional (see 
+   `examples/working/misc/cond_ddf`)
 
 # 0.3 (Nov 10, 2019)
 * Updated implementation of _broadcast_ nodes. Insertion of _broadcast_ node is now automatic when